Use a more concise way to fix compatibility issue for OutputWriter
